+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.0.2] - 15.10.2019.
 
 * Updated dependencies
+* Provide example
 
 ## [1.0.1] - 13.10.2019.